user controller logic

+++ app/controllers/users_controller.rb
@@ -1,7 +1,7 @@
-
 class UsersController < ApplicationController
-before_filter :authorize, only: [:edit, :update, :destroy]
 
+before_filter :authorize, only: [:edit, :update, :destroy]
+before_filter :get_user, except: [:new, :create, :index]
 
   def create
     @user = User.new(user_params)
@@ -26,36 +26,50 @@ before_filter :authorize, only: [:edit, :update, :destroy]
   end
 
   def show
-    @user = User.find(params[:id])
-    @toys = @user.toys.paginate(:page => params[:page], :per_page => 6)
-    @cities = City.all
+    if current_user
+      @user = User.find(params[:id])
+      @toys = @user.toys.paginate(:page => params[:page], :per_page => 6)
+      @cities = City.all
+    else
+      redirect_to root_path
+    end
   end
 
   def edit
-    @user = current_user
+    if current_user
+      if current_user == @user
+        @user = current_user
+      else
+        redirect_to user_path(current_user)
+      end
+    else
+      redirect_to root_path
+    end
   end
 
   def update
-  @user = current_user
-  user_params = params.require(:user).permit(:f_name, :l_name, :email, :avatar, :password, :bio)
-  if @user.update_attributes(user_params)
-    redirect_to user_path(@user)
-    flash[:notice] = "Your profile was updated!"
-  else 
-    flash[:error] = @user.error.full_messages.join(", ")
-    redirect_to edit_user_path
-  end
+    if current_user == @user
+      @user = current_user
+      if @user.update_attributes(user_params)
+        redirect_to user_path(@user)
+        flash[:notice] = "Your profile was updated!"
+      else 
+        flash[:error] = @user.error.full_messages.join(", ")
+        redirect_to edit_user_path
+      end
+    end
   end
 
   def destroy
-    @user = User.find(params[:id])
-  if @user.destroy
-      flash[:notice] = "Accout Deleted"
-      session[:user_id] = nil
-      redirect_to root_path
-    else
-      flash[:error] = @user.error.full_messages.join(", ")
-      redirect_to edit_user_path(@user)
+    if current_user == @user
+      if @user.destroy
+        flash[:notice] = "Accout Deleted"
+        session[:user_id] = nil
+        redirect_to root_path
+      else
+        flash[:error] = @user.error.full_messages.join(", ")
+        redirect_to edit_user_path(@user)
+      end
     end
   end
 
@@ -64,5 +78,9 @@ before_filter :authorize, only: [:edit, :update, :destroy]
     # Never trust parameters from the scary internet, only allow the white list through.
     def user_params
       params.require(:user).permit(:f_name, :l_name, :email, :avatar, :password, :bio, :toy_show, :toy_id, :city_id)
+    end
+  
+    def get_user
+      @user = User.find_by_id(params[:id])
     end
 end
